+++ bootstrap/include/frost/collections/Array_types.h
@@ -3,7 +3,7 @@
 typedef struct frost$core$Class frost$core$Class;
 #include "frost/core/Int32_types.h"
 #include "frost/core/UInt8_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 typedef struct frost$core$Object frost$core$Object;
 typedef struct frost$core$String frost$core$String;
 
@@ -11,8 +11,8 @@ typedef struct frost$collections$Array {
     frost$core$Class* $class;
     frost$core$Int32 refCount;
     frost$core$UInt8 $flags;
-    frost$core$Int64 _count;
-    frost$core$Int64 capacity;
+    frost$core$Int _count;
+    frost$core$Int capacity;
     frost$core$Object** data;
 } frost$collections$Array;
 #include "frost/core/Class.h"
