+++ bootstrap/include/frost/collections/Iterator/RangeIterator_types.h
@@ -4,7 +4,7 @@ typedef struct frost$core$Class frost$core$Class;
 #include "frost/core/Int32_types.h"
 #include "frost/core/UInt8_types.h"
 typedef struct frost$collections$Iterator frost$collections$Iterator;
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 typedef struct frost$core$Object frost$core$Object;
 #include "frost/core/Bit_types.h"
 typedef struct frost$core$String frost$core$String;
@@ -14,9 +14,9 @@ typedef struct frost$collections$Iterator$RangeIterator {
     frost$core$Int32 refCount;
     frost$core$UInt8 $flags;
     frost$collections$Iterator* base;
-    frost$core$Int64 current;
-    frost$core$Int64$nullable end;
-    frost$core$Int64 step;
+    frost$core$Int current;
+    frost$core$Int$nullable end;
+    frost$core$Int step;
     frost$core$Object* pending;
     frost$core$Bit _done;
 } frost$collections$Iterator$RangeIterator;
